backend/requirements.txt
celery==5.2.7
redis==4.5.1
sqlalchemy==1.4.46
selenium==4.8.3
undetected-chromedriver==3.2.3
fake_useragent==1.1.3
moviepy==1.0.3
requests==2.28.2
tqdm==4.64.1
PySocks==1.7.1
python-dotenv==0.21.1
